requirements.txt
aiohttp==3.8.5
beautifulsoup4==4.12.2
colorama==0.4.6
colorlog==6.7.0
faiss_cpu==1.7.4
fastapi==0.104.1
fasttext_langdetect==1.0.5
langchain==0.0.341
nest_asyncio==1.5.8
numpy==1.26.2
openai==0.28.1
pandas==2.1.3
psutil==5.9.5
pydantic==2.5.2
PyMySQL==1.1.0
python-dotenv==1.0.0
regex==2023.10.3
Requests==2.31.0
scikit_learn==1.3.1
semanticscholar==0.5.0
streamlit==1.27.2
tiktoken==0.5.1
torch==2.1.0
tqdm==4.66.1
uvicorn[standard]==0.24.0.post1
PyMuPDF==1.23.6
python-multipart==0.0.6
matplotlib==3.8.2
plotly==5.18.0
SQLAlchemy==2.0.23
cryptography==41.0.7
aiomysql==0.2.0
websockets==12.0
wsproto==1.2.0
lxml==4.9.3
openpyxl==3.1.2
serpapi==0.1.5
google-search-results==2.4.2
python-docx==1.1.0
portalocker==2.8.2




